Move game_items::Lookup class definition to GameItems/Lookup.h
